+++ templates/base.py
@@ -3,6 +3,33 @@ from .fonts.font import font_5x8
 from PIL import ImageFont
 from PIL import Image
 
+class box:
+    def __init__(self, cord=(0,0), value="", h=8, w=8, fgcolor=bytearray(b'\xff\xff\xff'), bgcolor=bytearray(b'\x00\x20\x00')):
+        self.x = cord[0]
+        self.y = cord[1]
+        self.h = h
+        self.w = w
+        self.value = value
+        self.bgcolor = bgcolor
+        self.fgcolor = fgcolor
+    
+    def set_values(self, cord=(0,0), value="", fgcolor=None, bgcolor=None):
+        self.x = cord[0]
+        self.y = cord[1]
+        self.value = value
+        self.bgcolor = bgcolor if bgcolor else self.bgcolor
+        self.fgcolor = fgcolor if fgcolor else self.fgcolor
+
+    def __eq__(self, other):
+        return ( 
+            self.value == other.value and 
+            self.fgcolor == other.fgcolor and 
+            self.bgcolor == other.bgcolor 
+        )
+
+    def get_cord(self, xoffset=0, yoffset=0):
+        return (self.x + xoffset, self.y + yoffset) 
+
 class base:
     marquee = None
     
@@ -51,3 +78,33 @@ class base:
                     data += (x+x_offset).to_bytes(2,"big") + (y+y_offset).to_bytes(1,"big") 
                     data += bytearray(list(im.getpixel((x,y))))
                     self.process_raw(data)
+    
+    def draw_7seg_digit(self, number, x_offset=0, y_offset=0, 
+            fgcolor=bytearray(b'\xba\x99\x10'), bgcolor=bytearray(b'\x00\x00\x00')):
+        seven_seg = {
+            "a" : { "cord" : (x_offset, y_offset), "h" : 1, "w" : 5 },
+            "b" : { "cord" : (5+x_offset, y_offset), "h" : 6, "w" : 1 },
+            "c" : { "cord" : (5+x_offset, 6+y_offset), "h" : 6, "w" : 1 },
+            "d" : { "cord" : (x_offset, 12+y_offset), "h" : 1, "w" : 5 },
+            "e" : { "cord" : (x_offset, 6+y_offset), "h" : 6, "w" : 1 },
+            "f" : { "cord" : (x_offset, y_offset), "h" : 6, "w" : 1 },
+            "g" : { "cord" : (x_offset, 6+y_offset), "h" : 1, "w" : 5 },
+        }
+        digit = {
+            "1" : ["b", "c"],
+            "2" : ["a", "b", "g", "e", "d"],
+            "3" : ["a", "b", "c", "d", "g"],
+            "4" : ["b", "c", "f", "g"],
+            "5" : ["a", "c", "d", "f", "g"],
+            "6" : ["a", "c", "d", "e", "f", "g"],
+            "7" : ["a", "b", "c"],
+            "8" : ["a", "b", "c", "d", "e", "f", "g"],
+            "9" : ["a", "b", "c", "d", "f", "g"],
+            "0" : ["a", "b", "c", "d", "e", "f"],
+        }
+        for seg,params in seven_seg.items():
+            c = bgcolor
+            if seg in digit[str(number)[0]]:
+                c = fgcolor
+            self.draw_box(**seven_seg[seg], color=c)
+
